+++ code/GP/server/commands/handle_delprofile.cpp
@@ -30,7 +30,7 @@ namespace GP {
 		}
 		
 		s << "\\dpr\\" << (int)(error_details.response_code == TaskShared::WebErrorCode_Success);
-		s << "\\id\\" << (int)extra;
+		s << "\\id\\" << (ptrdiff_t)extra;
 		((GP::Peer *)peer)->SendPacket((const uint8_t *)s.str().c_str(),s.str().length());
 
 		if ((error_details.response_code == TaskShared::WebErrorCode_Success)) {
@@ -40,7 +40,7 @@ namespace GP {
 	void Peer::handle_delprofile(OS::KVReader data_parser) {
 		TaskShared::ProfileRequest request;
 		request.profile_search_details.id = m_profile.id;
-		request.extra = (void *)data_parser.GetValueInt("id");
+		request.extra = (void *)(ptrdiff_t)data_parser.GetValueInt("id");
 		request.peer = this;
 		request.peer->IncRef();
 		request.type = TaskShared::EProfileSearch_DeleteProfile;
