+++ code/peerchat/server/commands/handle_setckey.cpp
@@ -30,7 +30,7 @@ namespace Peerchat {
 			do_combine = true;
 		}
 		if (do_combine) {
-			for (int i = 6; i < data_parser.size(); i++) {
+			for (size_t i = 6; i < data_parser.size(); i++) {
 				kv_string = kv_string.append(" ").append(data_parser.at(i));
 			}
 		}
